Uninformed Search
a program to implement the following search algorithms, to help Alice
find optimal traversal path(s) to send her message to reach Noah.
1. Breadth-first search (30 pts)
2. Depth-first search (30 pts)
3. Uniform-cost search using time as cost (40 pts)
Note: When the costs of expanding two or more nodes are equal, we
require that you expand
them alphabetically (i.e. David before Sophie).

The program should output the nodes (separated by “-”) in the order that
show the
path of forwarding the message. For example,
Alice-Ben-Frank-Gill-Mark-Noah.
There should be three .txt output files (one for each algorithm).
1. breadth-first.result.txt
2. depth-first.result.txt
3. uniform-cost.result.txt